+++ app/Http/Controllers/API/Item/ItemController.php
@@ -23,6 +23,14 @@ class ItemController extends Controller
             }
 
             $item = \App\Models\Item::all();
+
+            if ($item->isEmpty()) {
+                return response()->json([
+                    'success' => false,
+                    'message' => 'Item not found',
+                ], 404);
+            }
+
             return response()->json([
                 'success' => true,
                 'message' => 'List All Item',
@@ -100,6 +108,14 @@ class ItemController extends Controller
             }
 
             $item = \App\Models\Item::find($id);
+
+            if (!$item) {
+                return response()->json([
+                    'success' => false,
+                    'message' => 'Item not found',
+                ], 404);
+            }
+
             return response()->json([
                 'success' => true,
                 'message' => 'Item retrieved successfully',
@@ -143,6 +159,14 @@ class ItemController extends Controller
             }
 
             $item = \App\Models\Item::find($id);
+            
+            if(!$item) {
+                return response()->json([
+                    'success' => false,
+                    'message' => 'Item not found',
+                ], 404);
+            }
+
             $item->categories_id = $request->categories_id;
             $item->name = $request->name;
             $item->price = $request->price;
@@ -176,6 +200,14 @@ class ItemController extends Controller
             }
 
             $item = \App\Models\Item::find($id);
+            
+            if(!$item) {
+                return response()->json([
+                    'success' => false,
+                    'message' => 'Item not found',
+                ], 404);
+            }
+
             $item->delete();
 
             return response()->json([
